+++ .env.sample
@@ -1,0 +1,14 @@
+API_NAME=nuxt-skeleton
+NODE_ENV=dev
+PORT=3000
+LOG_LEVEL=debug
+CLEAN_ARCHITECTURE_API=
+# OIDC MODULE CONFIG
+NUXT_OIDC_TOKEN_KEY=
+NUXT_OIDC_SESSION_SECRET=
+NUXT_OIDC_AUTH_SESSION_SECRET=
+# KEYCLOAK PROVIDER CONFIG
+NUXT_OIDC_PROVIDERS_KEYCLOAK_CLIENT_SECRET=
+NUXT_OIDC_PROVIDERS_KEYCLOAK_CLIENT_ID=
+NUXT_OIDC_PROVIDERS_KEYCLOAK_BASE_URL=http://localhost:8080/realms/playground
+NUXT_OIDC_PROVIDERS_KEYCLOAK_REDIRECT_URI=http://localhost:3000/auth/keycloak/callback
