--- src/middleware/auth.ts
@@ -1,6 +1,0 @@
-export default defineNuxtRouteMiddleware(async () => {
-  const { loggedIn } = useOidcAuth()
-  if(!loggedIn.value){
-    return navigateTo("/guest", {replace: true});
-  }
-});
