+++ src/pages/guest.vue
@@ -1,5 +1,5 @@
 <template>
   <div>
-      <h1>vista invitado</h1>
+    <h1>vista invitado</h1>
   </div>
 </template>
