+++ src/pages/index.vue
@@ -2,37 +2,36 @@
 const { loggedIn, user, refresh, login, logout, currentProvider, clear } =
   useOidcAuth();
 </script>
-
 <template>
   <div class="w-full grid grid-cols-2">
     <div class="col-start-1 flex flex-col gap-4 items-center">
       <p class="text-xl">Login with</p>
-      <button class="btn-base btn-login" @click="login()">
+      <v-btn class="btn-base btn-login" @click="login()">
         <span class="i-majesticons-login-line" />
         <span class="pl-2">Default provider</span>
-      </button>
+      </v-btn>
       <p>Logged in: {{ loggedIn }}</p>
       <p>Current provider: {{ currentProvider }}</p>
-      <button
+      <v-btn
         class="btn-base btn-login"
         :disabled="!loggedIn || !user.canRefresh"
         @click="refresh()"
       >
         <span class="i-majesticons-refresh" />
         <span class="pl-2">Refresh</span>
-      </button>
-      <button
+      </v-btn>
+      <v-btn
         class="btn-base btn-login"
         :disabled="!loggedIn"
         @click="logout(currentProvider)"
       >
         <span class="i-majesticons-logout-line" />
         <span class="pl-2">Logout</span>
-      </button>
-      <button class="btn-base btn-login" :disabled="!loggedIn" @click="clear()">
+      </v-btn>
+      <v-btn class="btn-base btn-login" :disabled="!loggedIn" @click="clear()">
         <span class="i-majesticons-delete-bin-line" />
         <span class="pl-2">Clear session</span>
-      </button>
+      </v-btn>
     </div>
     <div class="col-start-2">
       <p class="pb-4 text-xl">User object</p>
